refactor(editor): Update EmbedLinkBlock component and related files
