+++ apps/frontend/apps/lyra-app/components/editor/extensions/v1/Embed/view/embed-link/EmbedLinkBlock.vue
@@ -1,63 +1,75 @@
 <template>
   <NodeViewWrapper>
-    <Card class="relative h-auto flex bg-clip-border rounded-xl bg-white text-gray-700 shadow-md w-full dark:bg-gray-800 dark:text-gray-300">
-      <div v-if="props.node.attrs.layout === 'image-left'" class="flex flex-row">
-        <div class="relative w-2/5 m-0 overflow-hidden text-gray-700 bg-white rounded-r-none bg-clip-border rounded-xl shrink-0 dark:bg-gray-700">
-          <NuxtImg :src="props.node.attrs.image" :alt="props.node.attrs.title" class="object-cover w-full h-full" />
+    <div class="w-full max-w-4xl mx-auto">
+      <Card>
+        <div
+          :class="
+            cn(
+              'flex flex-col',
+              props.node.attrs.layout === 'right'
+                ? 'md:flex-row-reverse'
+                : props.node.attrs.layout === 'left'
+                ? 'md:flex-row'
+                : props.node.attrs.layout === 'top'
+                ? 'flex-col'
+                : ''
+            )
+          "
+        >
+          <div
+            :class="
+              cn(
+                'flex-grow justify-center items-center',
+                props.node.attrs.layout === 'top'
+                  ? 'w-full p-2'
+                  : 'w-full md:w-1/3 p-0 h-full'
+              )
+            "
+          >
+            <AspectRatio :ratio="16 / 9" class="bg-muted h-full">
+              <NuxtImg
+                :src="props.node.attrs.image"
+                :alt="props.node.attrs.title"
+                class="rounded-md object-cover"
+                fit="cover"
+                preload
+                loading="lazy"
+              />
+            </AspectRatio>
+          </div>
+          <div
+            :class="
+              cn(
+                props.node.attrs.layout === 'top'
+                  ? 'w-full'
+                  : 'w-full md:w-1/2'
+              )
+            "
+          >
+            <CardContent>
+              <h3 class="line-clamp-1">{{ props.node.attrs.title }}</h3>
+              <p class="line-clamp-2">{{ props.node.attrs.description }}</p>
+              <a
+                :href="props.node.attrs.url"
+                target="_blank"
+                rel="noopener noreferrer"
+                class="line-clamp-1 block text-xs font-sans hover:underline overflow-hidden"
+              >
+                {{ props.node.attrs.url }}
+              </a>
+            </CardContent>
+          </div>
         </div>
-        <div class="p-6">
-          <h5 class="line-clamp-1 block mb-2 font-sans text-xl antialiased font-semibold leading-snug tracking-normal text-blue-gray-900 dark:text-white">
-            {{ props.node.attrs.title }}
-          </h5>
-          <p class="line-clamp-2 block mb-8 font-sans text-base antialiased font-normal leading-relaxed text-gray-700 dark:text-gray-400">
-            {{ props.node.attrs.description }}
-          </p>
-          <a :href="url" target="_blank" rel="noopener noreferrer" class="line-clamp-1 block text-xs font-sans hover:underline overflow-hidden dark:text-blue-400">
-            {{ url }}
-          </a>
-        </div>
-      </div>
-      <div v-else-if="props.node.attrs.layout === 'image-right'" class="flex flex-row-reverse">
-        <div class="relative w-2/5 m-0 overflow-hidden text-gray-700 bg-white rounded-l-none bg-clip-border rounded-xl shrink-0 dark:bg-gray-700">
-          <NuxtImg :src="props.node.attrs.image" :alt="props.node.attrs.title" class="object-cover w-full h-full" />
-        </div>
-        <div class="p-6">
-          <h5 class="line-clamp-1 block mb-2 font-sans text-xl antialiased font-semibold leading-snug tracking-normal text-blue-gray-900 dark:text-white">
-            {{ props.node.attrs.title }}
-          </h5>
-          <p class="line-clamp-2 block mb-8 font-sans text-base antialiased font-normal leading-relaxed text-gray-700 dark:text-gray-400">
-            {{ props.node.attrs.description }}
-          </p>
-          <a :href="url" target="_blank" rel="noopener noreferrer" class="line-clamp-1 block text-xs font-sans hover:underline overflow-hidden dark:text-blue-400">
-            {{ url }}
-          </a>
-        </div>
-      </div>
-      <div v-else class="flex flex-col">
-        <div class="relative w-full m-0 overflow-hidden text-gray-700 bg-white rounded-b-none bg-clip-border rounded-xl shrink-0 dark:bg-gray-700">
-          <NuxtImg :src="props.node.attrs.image" :alt="props.node.attrs.title" class="object-cover w-full h-full" />
-        </div>
-        <div class="p-6">
-          <h5 class="line-clamp-1 block mb-2 font-sans text-xl antialiased font-semibold leading-snug tracking-normal text-blue-gray-900 dark:text-white">
-            {{ props.node.attrs.title }}
-          </h5>
-          <p class="line-clamp-2 block mb-8 font-sans text-base antialiased font-normal leading-relaxed text-gray-700 dark:text-gray-400">
-            {{ props.node.attrs.description }}
-          </p>
-          <a :href="url" target="_blank" rel="noopener noreferrer" class="line-clamp-1 block text-xs font-sans hover:underline overflow-hidden dark:text-blue-400">
-            {{ url }}
-          </a>
-        </div>
-      </div>
-    </Card>
+      </Card>
+    </div>
   </NodeViewWrapper>
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
+import { cn } from "~/lib/utils";
 import { nodeViewProps, NodeViewWrapper } from "@tiptap/vue-3";
-import { Card } from "~/components/ui/card";
+import { Card, CardContent } from "@/components/ui/card";
+import { AspectRatio } from "@/components/ui/aspect-ratio";
 
 const props = defineProps(nodeViewProps);
-const url = ref<string>(props.node.attrs.url);
 </script>
